refactor(ui/input): optimize input decorator display
