+++ packages/varlet-ui/src/form/__tests__/index.spec.js
@@ -149,7 +149,7 @@ test('test form with select', async () => {
 
   const triggerEvents = async () => {
     await wrapper.find('.var-chip--close').trigger('click')
-    await wrapper.find('.var-input-box__clear-icon').trigger('click')
+    await wrapper.find('.var-input-decorator__clear-icon').trigger('click')
     await wrapper.find('.var-select').trigger('click')
   }
