+++ packages/varlet-ui/src/input/input.less
@@ -33,7 +33,7 @@
     outline: none;
     border: none;
     background: transparent;
-    color: var(--input-box-input-text-color);
+    color: var(--input-decorator-input-text-color);
     caret-color: var(--input-focus-color);
     transition: color 0.25s;
     font: inherit;
@@ -44,11 +44,11 @@
   }
 
   &--disabled {
-    color: var(--input-box-disabled-color);
+    color: var(--input-decorator-disabled-color);
     cursor: not-allowed;
   }
 
   &--caret-error {
-    caret-color: var(--input-box-error-color);
+    caret-color: var(--input-decorator-error-color);
   }
 }
