--- packages/varlet-ui/src/input/inputBox.less
@@ -1,276 +1,0 @@
-:root {
-  --input-box-input-text-color: #555;
-  --input-box-error-color: var(--color-danger);
-  --input-box-blur-color: #888;
-  --input-box-focus-color: var(--color-primary);
-  --input-box-placeholder-size: 16px;
-  --input-box-padding-top: 16px;
-  --input-box-padding-bottom: 16px;
-  --input-box-padding-left: 14px;
-  --input-box-padding-right: 14px;
-  --input-box-icon-padding: 20px 0 0;
-  --input-box-icon-size: 20px;
-  --input-box-line-size: 1px;
-  --input-box-line-focus-size: 2px;
-  --input-box-disabled-color: var(--color-text-disabled);
-}
-
-.var-input-box {
-  width: 100%;
-  color: var(--input-box-input-text-color);
-
-  &__controller {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    position: relative;
-  }
-
-  &__wrap {
-    position: relative;
-    flex-grow: 1;
-    display: flex;
-    flex-direction: column;
-  }
-
-  &__icon {
-    display: flex;
-    align-items: center;
-    font-size: 20px;
-  }
-
-  &__placeholder {
-    max-width: 100%;
-    position: absolute;
-    top: calc(
-      var(--input-box-padding-top) +
-        (100% - var(--input-box-padding-top) - var(--input-box-padding-bottom) - var(--input-box-placeholder-size)) / 2
-    );
-    font-size: var(--input-box-placeholder-size);
-    color: var(--input-box-blur-color);
-    line-height: 1em;
-    pointer-events: none;
-    cursor: inherit;
-    transform: translate(0, 0);
-  }
-
-  &__clear-icon[var-input-box-cover] {
-    display: flex;
-    font-size: 15px;
-    margin-left: 4px;
-  }
-
-  &--non-hint {
-    padding-top: 0;
-  }
-
-  &--placeholder-hint {
-    font-size: 0.75em;
-  }
-
-  &--placeholder-hidden {
-    visibility: hidden;
-  }
-
-  &--focus {
-    color: var(--input-box-focus-color);
-  }
-
-  &--disabled {
-    color: var(--input-box-disabled-color);
-    cursor: not-allowed;
-  }
-
-  &--error {
-    color: var(--input-box-error-color);
-  }
-
-  &--standard {
-    --input-box-padding-top: 20px;
-    --input-box-padding-bottom: 4px;
-
-    .var-input-box__wrap {
-      padding: var(--input-box-padding-top) 0 var(--input-box-padding-bottom);
-    }
-
-    .var-input-box--wrap-non-hint {
-      padding: 4px 0;
-    }
-
-    .var-input-box__icon {
-      padding: var(--input-box-icon-padding);
-    }
-
-    .var-input-box__placeholder {
-      left: 0;
-      transform-origin: left;
-      transition: top 0.3s, transform 0.3s, font-size 0.3s, color 0.25s;
-    }
-
-    .var-input-box--placeholder-hint {
-      top: 0;
-      transform: translate(0, 0);
-    }
-
-    .var-input-box--placeholder-non-hint {
-      top: 50%;
-      transform: translate(0, -50%);
-    }
-
-    .var-input-box__line {
-      width: 100%;
-      height: var(--input-box-line-size);
-      background: var(--input-box-blur-color);
-      transition: background-color 0.25s;
-    }
-
-    .var-input-box--line-disabled {
-      background: var(--input-box-disabled-color);
-    }
-
-    .var-input-box__dot {
-      width: 100%;
-      height: var(--input-box-line-focus-size);
-      background: var(--input-box-focus-color);
-      transform: scaleX(0);
-      transform-origin: center;
-      transition: transform 0.3s var(--cubic-bezier), background-color 0.25s;
-    }
-
-    .var-input-box--line-focus {
-      transform: scaleX(1);
-    }
-
-    .var-input-box--line-error {
-      background: var(--input-box-error-color);
-    }
-  }
-
-  &--outlined {
-    .var-input-box__controller {
-      padding: 0 var(--input-box-padding-right) 0 var(--input-box-padding-left);
-      cursor: text;
-    }
-    .var-input-box__wrap {
-      padding: var(--input-box-padding-top) 0 var(--input-box-padding-bottom) 0;
-    }
-
-    .var-input-box__placeholder {
-      transform-origin: left;
-      transition: top 0.3s, transform 0.3s, font-size 0.3s, margin 0.3s, color 0.25s;
-    }
-
-    .var-input-box--placeholder-hint {
-      margin: 0 2px;
-      top: 0;
-      transform: translateY(-50%);
-    }
-
-    .var-input-box__line {
-      width: 100%;
-      height: 100%;
-      position: absolute;
-      top: 0;
-      left: 0;
-      display: flex;
-      align-items: center;
-      pointer-events: none;
-
-      &__start {
-        flex: 0 0 12px;
-        height: 100%;
-        border-radius: 4px 0 0 4px;
-        border: 0 solid var(--input-box-blur-color);
-        border-top-width: var(--input-box-line-size);
-        border-bottom-width: var(--input-box-line-size);
-        border-inline-start-width: var(--input-box-line-size);
-        transition: border 0.1s;
-      }
-
-      &__notch {
-        max-width: calc(100% - 24px);
-        height: 100%;
-        font: inherit;
-        border: 0 solid var(--input-box-blur-color);
-        border-top-width: var(--input-box-line-size);
-        border-bottom-width: var(--input-box-line-size);
-        transition: border 0.1s;
-      }
-
-      &__notch--hint {
-        border-top-width: 0 !important;
-      }
-
-      &__placeholder {
-        max-width: 100%;
-        visibility: hidden;
-        font-size: 0.75em;
-        margin: 0 4px;
-      }
-
-      &__end {
-        flex: 1;
-        height: 100%;
-        border-radius: 0 4px 4px 0;
-        border: 0 solid var(--input-box-blur-color);
-        border-top-width: var(--input-box-line-size);
-        border-bottom-width: var(--input-box-line-size);
-        border-inline-end-width: var(--input-box-line-size);
-        transition: border 0.1s;
-      }
-    }
-
-    .var-input-box--line-focus {
-      .var-input-box__line__start {
-        border: 0 solid var(--input-box-focus-color);
-        border-top-width: var(--input-box-line-focus-size);
-        border-bottom-width: var(--input-box-line-focus-size);
-        border-inline-start-width: var(--input-box-line-focus-size);
-      }
-
-      .var-input-box__line__notch {
-        border: 0 solid var(--input-box-focus-color);
-        border-top-width: var(--input-box-line-focus-size);
-        border-bottom-width: var(--input-box-line-focus-size);
-      }
-
-      .var-input-box__line__end {
-        border: 0 solid var(--input-box-focus-color);
-        border-top-width: var(--input-box-line-focus-size);
-        border-bottom-width: var(--input-box-line-focus-size);
-        border-inline-end-width: var(--input-box-line-focus-size);
-      }
-    }
-
-    .var-input-box--line-disabled {
-      border-radius: 4px;
-      border: var(--input-box-line-size) solid var(--input-box-disabled-color);
-    }
-
-    .var-input-box--line-error {
-      border-color: var(--input-box-error-color) !important;
-    }
-  }
-
-  &--small {
-    .var-input-box--standard& {
-      --input-box-padding-top: 16px;
-      --input-box-icon-padding: 16px 0 0;
-
-      .var-input-box__wrap {
-        padding: var(--input-box-padding-top) 0 var(--input-box-padding-bottom);
-      }
-    }
-
-    .var-input-box--outlined& {
-      --input-box-padding-top: 8px;
-      --input-box-padding-bottom: 8px;
-      --input-box-padding-left: 12px;
-      --input-box-padding-right: 12px;
-
-      .var-input-box--placeholder-hint {
-        margin: 0 4px;
-      }
-    }
-  }
-}
