+++ packages/varlet-ui/src/input/inputDecorator.less
@@ -1,0 +1,282 @@
+:root {
+  --input-decorator-input-text-color: #555;
+  --input-decorator-error-color: var(--color-danger);
+  --input-decorator-blur-color: #888;
+  --input-decorator-focus-color: var(--color-primary);
+  --input-decorator-placeholder-size: 16px;
+  --input-decorator-padding-top: 16px;
+  --input-decorator-padding-bottom: 16px;
+  --input-decorator-padding-left: 14px;
+  --input-decorator-padding-right: 14px;
+  --input-decorator-icon-padding: 16px 0 16px;
+  --input-decorator-icon-size: 20px;
+  --input-decorator-line-size: 1px;
+  --input-decorator-line-focus-size: 2px;
+  --input-decorator-disabled-color: var(--color-text-disabled);
+}
+
+.var-input-decorator {
+  width: 100%;
+  color: var(--input-decorator-input-text-color);
+
+  &__controller {
+    width: 100%;
+    display: flex;
+    position: relative;
+  }
+
+  &__wrap {
+    position: relative;
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+  }
+
+  &__icon {
+    display: flex;
+    align-items: center;
+    font-size: var(--input-decorator-icon-size);
+    padding: var(--input-decorator-icon-padding);
+  }
+
+  &__placeholder {
+    max-width: 100%;
+    position: absolute;
+    top: calc(
+      var(--input-decorator-padding-top) +
+        (
+          100% - var(--input-decorator-padding-top) - var(--input-decorator-padding-bottom) -
+            var(--input-decorator-placeholder-size)
+        ) / 2
+    );
+    font-size: var(--input-decorator-placeholder-size);
+    line-height: 1em;
+    color: var(--input-decorator-blur-color);
+    pointer-events: none;
+    cursor: inherit;
+    transform: translate(0, 0);
+    vertical-align: middle;
+  }
+
+  &__placeholder-textarea {
+    top: var(--input-decorator-padding-top);
+  }
+
+  &__clear-icon[var-input-decorator-cover] {
+    display: flex;
+    font-size: 15px;
+    margin-left: 4px;
+    cursor: pointer;
+  }
+
+  &--non-hint {
+    padding-top: 0;
+  }
+
+  &--placeholder-hint {
+    font-size: 0.75em;
+  }
+
+  &--placeholder-hidden {
+    visibility: hidden;
+  }
+
+  &--focus {
+    color: var(--input-decorator-focus-color);
+  }
+
+  &--disabled {
+    color: var(--input-decorator-disabled-color);
+    cursor: not-allowed;
+  }
+
+  &--error {
+    color: var(--input-decorator-error-color);
+  }
+
+  &--standard {
+    --input-decorator-padding-top: 20px;
+    --input-decorator-padding-bottom: 4px;
+    --input-decorator-icon-padding: 20px 0 4px;
+
+    .var-input-decorator__wrap {
+      padding: var(--input-decorator-padding-top) 0 var(--input-decorator-padding-bottom);
+    }
+
+    .var-input-decorator--wrap-non-hint {
+      padding: 4px 0;
+    }
+
+    .var-input-decorator__placeholder {
+      left: 0;
+      transform-origin: left;
+      transition: top 0.3s, transform 0.3s, font-size 0.3s, color 0.25s;
+    }
+
+    .var-input-decorator--placeholder-hint {
+      top: 0;
+      transform: translate(0, 0);
+    }
+
+    .var-input-decorator--placeholder-non-hint {
+      top: 50%;
+      transform: translate(0, -50%);
+    }
+
+    .var-input-decorator__line {
+      width: 100%;
+      height: var(--input-decorator-line-size);
+      background: var(--input-decorator-blur-color);
+      transition: background-color 0.25s;
+    }
+
+    .var-input-decorator--line-disabled {
+      background: var(--input-decorator-disabled-color);
+    }
+
+    .var-input-decorator__dot {
+      width: 100%;
+      height: var(--input-decorator-line-focus-size);
+      background: var(--input-decorator-focus-color);
+      transform: scaleX(0);
+      transform-origin: center;
+      transition: transform 0.3s var(--cubic-bezier), background-color 0.25s;
+    }
+
+    .var-input-decorator--line-focus {
+      transform: scaleX(1);
+    }
+
+    .var-input-decorator--line-error {
+      background: var(--input-decorator-error-color);
+    }
+  }
+
+  &--outlined {
+    .var-input-decorator__controller {
+      padding: 0 var(--input-decorator-padding-right) 0 var(--input-decorator-padding-left);
+      cursor: text;
+    }
+    .var-input-decorator__wrap {
+      padding: var(--input-decorator-padding-top) 0 var(--input-decorator-padding-bottom) 0;
+    }
+
+    .var-input-decorator__placeholder {
+      transform-origin: left;
+      transition: top 0.3s, transform 0.3s, font-size 0.3s, margin 0.3s, color 0.25s;
+    }
+
+    .var-input-decorator--placeholder-hint {
+      margin: 0 2px;
+      top: 0;
+      transform: translateY(-50%);
+    }
+
+    .var-input-decorator__line {
+      width: 100%;
+      height: 100%;
+      position: absolute;
+      top: 0;
+      left: 0;
+      display: flex;
+      align-items: center;
+      pointer-events: none;
+
+      &__start {
+        flex: 0 0 12px;
+        height: 100%;
+        border-radius: 4px 0 0 4px;
+        border: 0 solid var(--input-decorator-blur-color);
+        border-top-width: var(--input-decorator-line-size);
+        border-bottom-width: var(--input-decorator-line-size);
+        border-inline-start-width: var(--input-decorator-line-size);
+        transition: border 0.1s;
+      }
+
+      &__notch {
+        max-width: calc(100% - 24px);
+        height: 100%;
+        font: inherit;
+        border: 0 solid var(--input-decorator-blur-color);
+        border-top-width: var(--input-decorator-line-size);
+        border-bottom-width: var(--input-decorator-line-size);
+        transition: border 0.1s;
+      }
+
+      &__notch--hint {
+        border-top-width: 0 !important;
+      }
+
+      &__placeholder {
+        max-width: 100%;
+        visibility: hidden;
+        font-size: 0.75em;
+        margin: 0 4px;
+      }
+
+      &__end {
+        flex: 1;
+        height: 100%;
+        border-radius: 0 4px 4px 0;
+        border: 0 solid var(--input-decorator-blur-color);
+        border-top-width: var(--input-decorator-line-size);
+        border-bottom-width: var(--input-decorator-line-size);
+        border-inline-end-width: var(--input-decorator-line-size);
+        transition: border 0.1s;
+      }
+    }
+
+    .var-input-decorator--line-focus {
+      .var-input-decorator__line__start {
+        border: 0 solid var(--input-decorator-focus-color);
+        border-top-width: var(--input-decorator-line-focus-size);
+        border-bottom-width: var(--input-decorator-line-focus-size);
+        border-inline-start-width: var(--input-decorator-line-focus-size);
+      }
+
+      .var-input-decorator__line__notch {
+        border: 0 solid var(--input-decorator-focus-color);
+        border-top-width: var(--input-decorator-line-focus-size);
+        border-bottom-width: var(--input-decorator-line-focus-size);
+      }
+
+      .var-input-decorator__line__end {
+        border: 0 solid var(--input-decorator-focus-color);
+        border-top-width: var(--input-decorator-line-focus-size);
+        border-bottom-width: var(--input-decorator-line-focus-size);
+        border-inline-end-width: var(--input-decorator-line-focus-size);
+      }
+    }
+
+    .var-input-decorator--line-disabled {
+      border-radius: 4px;
+      border: var(--input-decorator-line-size) solid var(--input-decorator-disabled-color);
+    }
+
+    .var-input-decorator--line-error {
+      border-color: var(--input-decorator-error-color) !important;
+    }
+  }
+
+  &--small {
+    .var-input-decorator--standard& {
+      --input-decorator-padding-top: 16px;
+      --input-decorator-icon-padding: 16px 0 0;
+
+      .var-input-decorator__wrap {
+        padding: var(--input-decorator-padding-top) 0 var(--input-decorator-padding-bottom);
+      }
+    }
+
+    .var-input-decorator--outlined& {
+      --input-decorator-padding-top: 8px;
+      --input-decorator-padding-bottom: 8px;
+      --input-decorator-padding-left: 12px;
+      --input-decorator-padding-right: 12px;
+
+      .var-input-decorator--placeholder-hint {
+        margin: 0 4px;
+      }
+    }
+  }
+}
