+++ packages/varlet-ui/src/select/select.less
@@ -55,7 +55,7 @@
     outline: none;
     border: none;
     font-size: inherit;
-    color: var(--input-box-input-text-color);
+    color: var(--input-decorator-input-text-color);
     word-break: break-all;
     overflow-x: hidden;
     transition: color 0.25s;
@@ -67,12 +67,12 @@
   }
 
   &--disabled {
-    color: var(--input-box-disabled-color);
+    color: var(--input-decorator-disabled-color);
     cursor: not-allowed;
   }
 
   &--error {
-    color: var(--input-box-error-color);
+    color: var(--input-decorator-error-color);
   }
 
   &--arrow-rotate {
